Replace celebrityninjas.com by api-ninjas.com
The former didn't really work during the cdCon 2022 period
(almost only 504 timeouts errors), the later is almost the same API
but was working correctly.

There are also some other fixes, such as send()-ing the actual
incident to Zeebe when it fails, plus some debugging information.

We hacked quite a bit on it during cdCon 2022 to make it work
correctly :)